Minor bug fix
Program could accept a shorter PAM site if it was contained within one
of the sequences within the degenerate possibilities of a stored site.
This would lead to serious problems.  Program now checks first to
ensure that stored and target PAM are the same length.  Also put in an
exception handler for when there are no perfect matches to a sequence
to warn user and display information.